Split profile page Add SharedProfile page
