+++ src/pagePartials/profile/created/BadgesCreatedSection.tsx
@@ -1,0 +1,10 @@
+import CreateNewBadge from '@/src/pagePartials/profile/created/CreateNewBadge'
+
+export default function BadgesCreatedSection() {
+  return (
+    <>
+      <h2>BadgesCreatedSection</h2>
+      <CreateNewBadge />
+    </>
+  )
+}
