+++ src/pagePartials/profile/inReview/BadgesInReviewSection.tsx
@@ -1,0 +1,10 @@
+import BadgesToCurate from '@/src/pagePartials/profile/inReview/BadgesToCurate'
+
+export default function BadgesInReviewSection() {
+  return (
+    <>
+      <h2>BadgesInReviewSection</h2>
+      <BadgesToCurate />
+    </>
+  )
+}
