+++ src/pagePartials/profile/myProfile/Pending.tsx
@@ -1,31 +1,31 @@
-import HistoryOutlinedIcon from '@mui/icons-material/HistoryOutlined'
+import NotificationsNoneOutlinedIcon from '@mui/icons-material/NotificationsNoneOutlined'
 import { Box, Divider, Typography } from '@mui/material'
 import { colors } from 'thebadge-ui-library'
 
 import SafeSuspense from '@/src/components/helpers/SafeSuspense'
 import { SectionTitleBox } from '@/src/pagePartials/home/SectionBoxes'
-import NearToExpireList from '@/src/pagePartials/profile/NearToExpireList'
+import PendingList from '@/src/pagePartials/profile/myProfile/PendingList'
 
-export default function NearToExpire() {
+export default function Pending() {
   return (
     <Box sx={{ width: '100%', px: 4, py: 2 }}>
       <SectionTitleBox>
         <Box sx={{ display: 'flex', alignItems: 'center', gap: 1, mb: 2 }}>
-          <HistoryOutlinedIcon color="purple" />
+          <NotificationsNoneOutlinedIcon color="green" />
           <Typography
-            color={colors.purple}
+            color={colors.green}
             fontWeight={900}
             lineHeight={'30px'}
             textAlign="center"
             variant={'h5'}
           >
-            Near to expire
+            Pending
           </Typography>
         </Box>
       </SectionTitleBox>
-      <Divider />
+      <Divider color={colors.green} />
       <SafeSuspense>
-        <NearToExpireList />
+        <PendingList />
       </SafeSuspense>
     </Box>
   )
